enclave-runc-wrapper: Add runc wrapper as a bridge with k8s
This change adds the runc wrapper which has a purpose of applying
policies set through Kubernetes namespaces. For now, it uses the
namespace labels proposed in the PSP Replacement KEP[0].

The runc wrapper checks whether the container which is about to be
started was scheduled by kubelet (by checking runc annotations set by
kubelet / runtime servers). If some non-default policy is set, that
policy is written for the particilar container to the BPF map, so then
BPF programs can be aware of that policy.

For now, there is no way of proving that those annotations really come
from Kubernetes. Coming up with some sane way of securely proving that
will be something to implement in a follow up work.

[0] https://github.com/kubernetes/enhancements/pull/2582

Fixes: #3
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>