bpf: Do not register containers
Since the runc wrapper got introduced, it's more convenient to register
new containers through it, instead of detecting then through BPF
programs.

This change also fixes the way how PID is determined for each process.

Another change is denying to launch any runc processes which are not
children of the runc wrapper.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>