Fix needless_borrow error, adjust to the Rust update
The newest Rust and clippy were reporting a couple needless_borrow errors.
This change fixes them.

The mistake was using `&` on variables which were already of type
`&[u8]`. So we were in fact passing `&&[u8]` to function calls.

There is also a change in the BPF skeleton (also due to the new Rust
version).

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>