docker: Mount the source tree during build
Before this change, containerized build was not updating the
lockc.skel.rs file on the host. This change fixes that by mounting the
source directory and ditching the mechanism of copying binaries from an
OCI atrifact.

Since the whole source is mounted, that means it gets all the files
produced by Meson during build. Therefore there is no need for custom
hacks for installation after a containerized build. We can simply do
`meson install --no-rebuild`. That's why this change simplifies the
`install-lockc.sh` script (executed by cloud-init in a VM) to always use
Meson.

To ensure that the newest Meson and Ninja are always used, install them
with pip.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>