bpf: Add mount policies
This change introduces the mount policies based on the `sb_mount` LSM
hook. That hook is used to monitor bind mounts performed by the
container runtime (runc) and check whether we should allow it.

The behavior for each policy:
* restricted: allow only bind mounts from sources defined in the
  `allowed_paths_restricted` BPF map
* baseline: allow only bind mounts from sources defined in the
  `allow_paths_baseline` BPF map
* privileged: allow all bind mounts

BPF programs don't predefine the content of the maps. That is going to
be done by lockcd in userspace.

Fixes: #4
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>