lockc: Add mount policies
This change uses the new BPF program based on `sb_mount` in lockcd, the
userspace part. lockcd is responsible for filling up the BPF maps with
directories which are allowed to be bind mounted into restricted and
baseline containers.

There are two configurable lists of directories to allow - for
restricted and for baseline containers.

By default, the list for restricted containers contains only directories
which container engines (docker, containerd, libpod/podman/cri-o) use:

* /dev/pts - Pseudo-Terminal Device
* /var/lib/containers/storage - filesystem layers used by libpod
* /var/run/container - filesystem layers used by containerd

The list for baseline containers contains those additional dirs which
can be used to bind mount data files from host:

* /home - for home content
* /var/data - for application data

This commit also brings the following changes in code:

* use bindgen to autogenerate Rust types and variables from C headers,
  to avoid having to define structs (especially those used in BPF maps)
  twice
* move the code responsible for BPF map operations and converting
  data to bytes into a dedicated trait and function, to avoid code
  duplication
* handle the --console-socket argument in runc - not doing so was
  causing occasional crashes of the wrapper and inconsistency in BPF
  maps

This change can be tested locally by running lockcd:

  sudo ./build/src/lockcd

And then running a container using wrapper as a runtime which tries to
mount an insecure, not allowed directory:

  podman --runtime $(pwd)/build/src/lockc-runc-wrapper run -ti -v /:/rootfs --rm registry.opensuse.org/opensuse/toolbox:latest

That command should **not** work.

On the other hand, running a container without -v or with bind mounts
from /home or /var/data should work.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>