Remove meson, use only cargo and dapper
This change removes meson buildsystem usage. It doesn't provide any
cargo integration, so we were pretty much gluing it with cargo via a
bash script. The only two features of meson we ended up using were:

* templating files
* installation

We came to conclusion that using a complex build system just for those
two tasks is not really worth it.

In this change, we achieve those two tasks by creating a custom xtask
for system-wide installation of binaries, config files and systemd
units. That `install` xtask is supposed to bring a similar behavior to
traditional `make install`.

The other two changes made are:

* removing BPF skeleton from the repo
* running tests as root

Fixes: #59
Fixes: #60
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>