bpf: Hide directories from access
This change adds new BPF programs which restrict the filesystem access
(open) inside containers based on configurable allow and deny BPF maps.
Allow/deny lists of directories are available in the usespace in lockc
configuration.

It can be tested by trying to access a denied directory in a container,
like:

  lockc-control-plane-0:~ # docker run --rm -it busybox sh
  / # ls /sys
  ls: can't open '/sys': Operation not permitted
  / # ls /home
  / #

Fixes: #48
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>