bpf: Adjust PID-related BPF map sizes to the current pid_max
Before this change, we assumed the worst case scenario about the number
of processes which can be schedules - the upper configurable limit for
x86_64 - which in theory is 4194304.

Allocating such a map size resulted in huge memory usage by lockc BPF
programs, over 500 MB, which is unacceptable for nodes with <=2 GB RAM
(which is usual both in IoT/ARM world and public clouds).

This change reads the currently configured limit by reading the
kernel.pid_max sysctl. On the most of systems it's configured to 32768.
That's way less than the maximum worst case scenario.

Memory usage before this change:

  lockc-control-plane-0:~ # free -m
                total        used        free      shared  buff/cache   available
  Mem:           1968         811         689           8         467        1011
  Swap:             0           0           0
  lockc-control-plane-0:~ # systemctl start lockcd
  lockc-control-plane-0:~ # free -m
                total        used        free      shared  buff/cache   available
  Mem:           1968        1420         154           8         393         339
  Swap:             0           0           0

After this change:

  lockc-control-plane-0:~ # free -m
                total        used        free      shared  buff/cache   available
  Mem:           1968         793         220           3         955        1037
  Swap:             0           0           0
  lockc-control-plane-0:~ # systemctl start lockcd
  lockc-control-plane-0:~ # free -m
                total        used        free      shared  buff/cache   available
  Mem:           1968         803         210           3         955        1028
  Swap:             0           0

Fixes: #82
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>