Ensure that the binary is named lockcd
In pull request #142 we made a mistake of converting lockc into full
binary crate, therefore renaming lockcd to lockc and removing a
possibility of creating multiple binaries (if there will be a need -
i.e. for some CLI tool).

This change fixes that by creating a bin/ directory with binaries again.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>