docs: Make a separate section for developers
This change makes a separate section for developers and moves
information about building from source and terraform. From now,
"Getting started" section is meant for operators and all dev-related
stuff should not be included there.

It also updates information about new location of Helm charts and
necessity of using a separate repo for using them (as a developer).

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>