tokio: Fix the way of using oneshot::Sender
According to docs of tokio::sync::oneshot::Sender.send, that method
"attempts to send a value on this channel, returning it back if it could
not be sent."

Err case of send() is just simply returning the sent value back. Quite
confusing, I was expecting some real error there. But well, according to
the example snippet:

https://docs.rs/tokio/1.16.1/tokio/sync/oneshot/struct.Sender.html#examples-1

we are just logging a custom error there.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>